requirements.txt
# Core
numpy>=1.23
scipy>=1.9
pandas>=1.5
matplotlib>=3.7
scikit-learn>=1.2
tqdm>=4.65
pyyaml>=6.0
joblib>=1.3

# Vision
opencv-python>=4.8
Pillow>=9.0
imgaug>=0.4

# Deep Learning
torch>=1.13
torchvision>=0.14
timm>=0.9
efficientnet_pytorch>=0.7

# Utilities
einops>=0.6
streamlit>=1.28